Add default stdout/err communicator
