+++ renku/cli/utils.py
@@ -20,20 +20,16 @@ r"""Communicator class for printing click output."""
 from functools import wraps
 
 import click
-from tqdm import tqdm
 
 import renku.core.utils.communication as communication
 
 
-class ClickCallback(communication.CommunicationCallback):
+class ClickCallback(communication.StandardOutput):
     """CommunicationCallback implementation for ``click`` messages."""
 
     INFO = click.style('Info: ', bold=True, fg='blue')
     WARNING = click.style('Warning: ', bold=True, fg='yellow')
     ERROR = click.style('Error: ', bold=True, fg='red')
-
-    progressbars = {}
-    progress_types = ['download']
 
     def echo(self, msg):
         """Write a message."""
@@ -54,43 +50,6 @@ class ClickCallback(communication.CommunicationCallback):
     def confirm(self, msg, abort=False):
         """Get confirmation for an action using a prompt."""
         return click.confirm(msg, abort=abort)
-
-    def start_progress(self, name, total, **kwargs):
-        """Start a new tqdm progressbar."""
-        if name in self.progressbars:
-            raise ValueError(
-                'Name {} is already a registered progressbar.'.format(name)
-            )
-
-        if 'type' not in kwargs:
-            kwargs['type'] = 'download'
-
-        if kwargs['type'] not in self.progress_types:
-            self.progressbars[name] = None
-        elif kwargs['type'] == 'download':
-            self.progressbars[name] = tqdm(
-                total=total,
-                unit='iB',
-                unit_scale=True,
-                desc=name,
-                leave=False,
-                bar_format='{desc:.32}: {percentage:3.0f}%|{bar}{r_bar}'
-            )
-
-    def update_progress(self, name, amount):
-        """Update a progressbar."""
-        if name not in self.progressbars or not self.progressbars[name]:
-            return
-
-        self.progressbars[name].update(amount)
-
-    def finalize_progress(self, name):
-        """End a progressbar."""
-        if name not in self.progressbars or not self.progressbars[name]:
-            return
-
-        self.progressbars[name].close()
-        del self.progressbars[name]
 
 
 def click_callback_communication(func):
